Fix for LOGBACK-894 cleanByPeriodOffset add time check

+++ logback-core/src/main/java/ch/qos/logback/core/rolling/helper/SizeAndTimeBasedArchiveRemover.java
@@ -39,7 +39,12 @@ public class SizeAndTimeBasedArchiveRemover extends DefaultArchiveRemover {
         parentDir, stemRegex);
 
     for (File f : matchingFileArray) {
-      f.delete();
+      Date fileLastModified = rc.getRelativeDate(new Date(f.lastModified()), -1);
+
+      if (fileLastModified.compareTo(dateOfPeriodToClean) <= 0) {
+        addInfo("deleting " + f);
+        f.delete();
+      }
     }
 
     if (parentClean) {
